requirements.txt
asn1crypto==0.24.0
attrs==17.4.0
Automat==0.6.0
blessings==1.7
certifi==2018.1.18
cffi==1.14.0
chardet==3.0.4
click==6.7
colorama==0.3.7
configobj==5.0.6
constantly==15.1.0
cryptography==2.1.4
cycler==0.11.0
Cython==0.29.20
dataclasses==0.8
fasteners==0.15
glfw==1.11.2
gpustat==0.6.0
httplib2==0.9.2
hyperlink==17.3.1
icdiff==1.9.1
idna==2.6
imageio==2.8.0
importlib-resources==5.4.0
incremental==16.10.1
joblib==1.1.1
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.3.1
macaroonbakery==1.1.3
matplotlib==3.3.4
monotonic==1.5
netifaces==0.10.4
numpy==1.19.5
nvidia-ml-py3==7.352.0
pandas==1.1.5
Pillow==7.2.0
protobuf==3.0.0
psutil==5.6.3
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycparser==2.20
pycrypto==2.6.1
pycurl==7.43.0.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyOpenSSL==17.5.0
pyparsing==3.0.9
pyRFC3339==1.0
pyserial==3.4
python-dateutil==2.8.2
python-debian==0.1.32
pytz==2018.3
pyxdg==0.25
PyYAML==3.12
requests==2.18.4
requests-unixsocket==0.1.5
scikit-learn==0.24.2
scipy==1.5.4
SecretStorage==2.3.1
service-identity==16.0.0
six==1.12.0
sklearn==0.0.post1
ssh-import-id==5.7
threadpoolctl==3.1.0
torch==1.10.1
torchvision==0.11.2
tqdm==4.64.1
trash==0.1.0
Twisted==17.9.0
typing-extensions==4.1.1
urllib3==1.22

